+++ Tests/Mocks/MockUser.swift
@@ -22,28 +22,10 @@
  SOFTWARE.
  */
 
-import Quick
-import Nimble
+import Foundation
 @testable import MessageKit
 
-final class SenderSpec: QuickSpec {
-
-    override func spec() {
-        describe("equality between two Senders") {
-            context("they have the same id ") {
-                it("should be equal") {
-                    let sender1 = MockUser(id: "1", displayName: "Steven")
-                    let sender2 = MockUser(id: "1", displayName: "Nathan")
-                    expect(sender1.id == sender2.id).to(equal(true))
-                }
-            }
-            context("they have a different id") {
-                it("should not be equal") {
-                    let sender1 = MockUser(id: "1", displayName: "Steven")
-                    let sender2 = MockUser(id: "2", displayName: "Nathan")
-                    expect(sender1.id == sender2.id).to(equal(false))
-                }
-            }
-        }
-    }
+struct MockUser: SenderType {
+    var id: String
+    var displayName: String
 }